split bit_cast
