+++ include/eve/detail/function/simd/arm/neon/movemask.hpp
@@ -9,7 +9,7 @@
 
 #include <eve/arch/logical.hpp>
 #include <eve/detail/meta.hpp>
-#include <eve/detail/function/bit_cast.hpp>
+#include <eve/module/core/regular/bit_cast.hpp>
 #include <eve/module/core/regular/convert.hpp>
 
 namespace eve::detail
