+++ include/eve/detail/function/simd/common/bit_cast.hpp
@@ -12,8 +12,8 @@
 
 namespace eve::detail
 {
-  template<callable_options O, typename T, typename Target>
-  EVE_FORCEINLINE constexpr auto bit_cast_(EVE_REQUIRES(cpu_), O const&, T const &a, as<Target> const &) noexcept
+  template<typename T, typename Target>
+  EVE_FORCEINLINE constexpr auto bit_cast_impl(cpu_, T const &a, as<Target>) noexcept
   {
     // Fixes bad codegen on some compilers
     if constexpr(std::is_same_v<T, Target>) return a;
